backend/requirements.txt
# FastAPI and ASGI Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlmodel>=0.0.22
psycopg2-binary>=2.9.10
alembic>=1.14.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.17

# Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.2.0

# Environment Variables
python-dotenv>=1.0.1

# Development
ruff>=0.8.0

# Phase III - AI Chatbot Dependencies
openai>=1.0.0
mcp>=0.1.0
